# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "parameters": [
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/paginated/all": {
      "get": {
        "operationId": "ProjectsController_getAllWithPagination",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/data/{uuid}": {
      "get": {
        "operationId": "ProjectsController_findByUuid",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/assign": {
      "post": {
        "operationId": "ProjectsController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectAssignment"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/add/comments": {
      "patch": {
        "operationId": "ProjectsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectComment"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/user/{userId}": {
      "get": {
        "operationId": "ProjectsController_getUserProjects",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/staff/all/{staffId}": {
      "get": {
        "operationId": "ProjectsController_getProjectsByStaff",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/projects/{projectId}/staffs": {
      "get": {
        "operationId": "ProjectsController_getProjectStaffs",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/task/project/{projectId}": {
      "get": {
        "operationId": "TaskController_findByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "day",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "urgency",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
...
    "/task/staff/{staffId}/project/{projectId}": {
      "get": {
        "operationId": "TaskController_getStaffTasksByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "day",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "urgency",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "info": {
    "title": "CN Squad API",
    "description": "The CN Squad API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StaffAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "city",
          "state",
          "country",
          "createdAt",
          "updatedAt"
        ]
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nameAbrv": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nameAbrv",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "StaffEmployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "employeeCode": {
            "type": "string"
          },
          "jobTitle": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "Full-Time",
              "Part-Time",
              "Contract",
              "Intern"
            ]
          },
          "workMode": {
            "type": "string",
            "enum": [
              "Onsite",
              "Remote",
              "Hybrid"
            ]
          },
          "hireDate": {
            "type": "string"
          },
          "reportingManager": {
            "type": "string"
          },
          "directReport": {
            "type": "string"
          },
          "contractEndDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended",
              "Terminated"
            ]
          },
          "workLocation": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "departmentalRole": {
            "$ref": "#/components/schemas/DepartmentalRole"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "employeeCode",
          "jobTitle",
          "employmentType",
          "workMode",
          "hireDate",
          "reportingManager",
          "directReport",
          "status",
          "department",
          "departmentalRole",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hobbies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "Male",
              "Female",
              "Other"
            ],
            "type": "string"
          },
          "maritalStatus": {
            "enum": [
              "Single",
              "Married",
              "Divorced",
              "Widowed"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/StaffAddress"
          },
          "employment": {
            "$ref": "#/components/schemas/StaffEmployment"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "resetToken": {
            "type": "string"
          },
          "resetTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "registrationToken": {
            "type": "string"
          },
          "registrationTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "isRegistered": {
            "type": "boolean"
          },
          "lastIpAddress": {
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isCeo": {
            "type": "boolean"
          },
          "isOnline": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "firstName",
          "lastName",
          "description",
          "hobbies",
          "dateOfBirth",
          "gender",
          "email",
          "phone",
          "password",
          "address",
          "employment",
          "roles",
          "isRegistered",
          "isCeo",
          "isOnline",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProjectAssignment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "project",
          "staff",
          "role"
        ]
      },
      "ProjectComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project",
          "staff",
          "text",
          "createdAt"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "timeLine": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "status": {
            "enum": [
              "not_started",
              "in_progress",
              "testing",
              "completed"
            ],
            "type": "string"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "priority": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "type": "string"
          },
          "prodUrl": {
            "type": "string"
          },
          "stagingUrl": {
            "type": "string"
          },
          "apk": {
            "type": "string"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/Staff"
          },
          "assignedTo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectAssignment"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectComment"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "projectName",
          "department",
          "status",
          "tasks",
          "priority",
          "startDate",
          "endDate",
          "createdBy",
          "assignedTo",
          "comments",
          "createdAt",
          "updatedAt"
        ]
      },
      "TaskAssignment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "task",
          "staff",
          "role"
        ]
      },
      "TaskComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          },
          "mentionedStaff": {
            "$ref": "#/components/schemas/Staff"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "text": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task",
          "mentionedStaff",
          "staff",
          "text",
          "createdAt",
          "updatedAt"
        ]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "taskModule": {
            "type": "string"
          },
          "sprint": {
            "type": "number"
          },
          "status": {
            "enum": [
              "Not_started",
              "In_progress",
              "Ready_To_Test",
              "Testing_In_Progress",
              "On_Hold",
              "Failed_Test",
              "Passed_Test",
              "Dev_Completed",
              "Completed"
            ],
            "type": "string"
          },
          "priority": {
            "enum": [
              "Low",
              "Medium",
              "High"
            ],
            "type": "string"
          },
          "urgency": {
            "enum": [
              "Short_Term",
              "Med_Term",
              "High_Term",
              "Immediately"
            ],
            "type": "string"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string"
          },
          "timeline": {
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "linkedProjectId": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Staff"
          },
          "assignedTo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskAssignment"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskComment"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "completed_Date": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "sprint",
          "status",
          "priority",
          "urgency",
          "createdBy",
          "assignedTo",
          "comments",
          "createdAt",
          "completed_Date",
          "updatedAt"
        ]
      },
      "SingleAssignmentDto": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "number"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "staffId",
          "role"
        ]
      },
      "AssignStaffDto": {
        "type": "object",
        "properties": {
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SingleAssignmentDto"
            }
          }
        },
        "required": [
          "assignments"
        ]
      },
      "AddCommentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "mentionedStaffId": {
            "type": "number"
          }
        },
        "required": [
          "text"
        ]
      },
...
    }
  }
}